back-end/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.5,<3.0
PyJWT>=2.9,<3.0
python-multipart>=0.0.9,<1.0
numpy>=1.26,<3.0
opencv-python>=4.10,<5.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0
